+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/magical/EnchantmentRune.java
@@ -3,7 +3,6 @@ package io.github.thebusybiscuit.slimefun4.implementation.items.magical;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.EnumMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -20,12 +19,12 @@ import org.bukkit.entity.Player;
 import org.bukkit.event.player.PlayerDropItemEvent;
 import org.bukkit.inventory.ItemStack;
 
+import io.github.thebusybiscuit.slimefun4.core.handlers.ItemDropHandler;
+import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SimpleSlimefunItem;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
-import me.mrCookieSlime.Slimefun.Objects.handlers.ItemDropHandler;
-import me.mrCookieSlime.Slimefun.SlimefunPlugin;
 import me.mrCookieSlime.Slimefun.api.Slimefun;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
@@ -49,10 +48,17 @@ public class EnchantmentRune extends SimpleSlimefunItem<ItemDropHandler> {
 
         for (Material mat : Material.values()) {
             List<Enchantment> enchantments = new ArrayList<>();
+
             for (Enchantment enchantment : Enchantment.values()) {
-                if (enchantment == Enchantment.BINDING_CURSE || enchantment == Enchantment.VANISHING_CURSE) continue;
-                if (enchantment.canEnchantItem(new ItemStack(mat))) enchantments.add(enchantment);
+                if (enchantment == Enchantment.BINDING_CURSE || enchantment == Enchantment.VANISHING_CURSE) {
+                    continue;
+                }
+
+                if (enchantment.canEnchantItem(new ItemStack(mat))) {
+                    enchantments.add(enchantment);
+                }
             }
+
             applicableEnchantments.put(mat, enchantments);
         }
     }
@@ -66,15 +72,16 @@ public class EnchantmentRune extends SimpleSlimefunItem<ItemDropHandler> {
                     return true;
                 }
 
-                Slimefun.runSync(() -> activate(p, e, item), 20L);
+                Slimefun.runSync(() -> addRandomEnchantment(p, e, item), 20L);
 
                 return true;
             }
+
             return false;
         };
     }
 
-    private void activate(Player p, PlayerDropItemEvent e, Item item) {
+    private void addRandomEnchantment(Player p, PlayerDropItemEvent e, Item item) {
         // Being sure the entity is still valid and not picked up or whatsoever.
         if (!item.isValid()) {
             return;
@@ -89,32 +96,36 @@ public class EnchantmentRune extends SimpleSlimefunItem<ItemDropHandler> {
             ItemStack target = entity.getItemStack();
 
             List<Enchantment> applicableEnchantmentList = applicableEnchantments.get(target.getType());
+
             if (applicableEnchantmentList == null) {
-                SlimefunPlugin.getLocal().sendMessage(p, "messages.enchantment-rune.fail", true);
+                SlimefunPlugin.getLocalization().sendMessage(p, "messages.enchantment-rune.fail", true);
                 return;
-            } else {
+            }
+            else {
                 applicableEnchantmentList = new ArrayList<>(applicableEnchantmentList);
             }
 
-            //Removing the enchantments that the item already has from enchantmentSet
-            for (Enchantment itemEnchantment : target.getEnchantments().keySet()) {
-                for (Enchantment applicableEnchantment : applicableEnchantmentList) {
-                    if (applicableEnchantment == itemEnchantment || applicableEnchantment.conflictsWith(itemEnchantment)) {
-                        applicableEnchantmentList.remove(applicableEnchantment);
+            // Removing the enchantments that the item already has from enchantmentSet
+            for (Enchantment enchantment : target.getEnchantments().keySet()) {
+                for (Enchantment possibleEnchantment : applicableEnchantmentList) {
+                    if (possibleEnchantment == enchantment || possibleEnchantment.conflictsWith(enchantment)) {
+                        applicableEnchantmentList.remove(possibleEnchantment);
                     }
                 }
             }
 
             if (applicableEnchantmentList.isEmpty()) {
-                SlimefunPlugin.getLocal().sendMessage(p, "messages.enchantment-rune.no-enchantment", true);
+                SlimefunPlugin.getLocalization().sendMessage(p, "messages.enchantment-rune.no-enchantment", true);
                 return;
             }
 
             Enchantment enchantment = applicableEnchantmentList.get(ThreadLocalRandom.current().nextInt(applicableEnchantmentList.size()));
             int level = 1;
+            
             if (enchantment.getMaxLevel() != 1) {
                 level = ThreadLocalRandom.current().nextInt(enchantment.getMaxLevel()) + 1;
             }
+            
             target.addEnchantment(enchantment, level);
 
             if (target.getAmount() == 1) {
@@ -134,11 +145,12 @@ public class EnchantmentRune extends SimpleSlimefunItem<ItemDropHandler> {
                         item.remove();
                         l.getWorld().dropItemNaturally(l, target);
 
-                        SlimefunPlugin.getLocal().sendMessage(p, "messages.enchantment-rune.success", true);
+                        SlimefunPlugin.getLocalization().sendMessage(p, "messages.enchantment-rune.success", true);
                     }
                 }, 10L);
-            } else {
-                SlimefunPlugin.getLocal().sendMessage(p, "messages.enchantment-rune.fail", true);
+            }
+            else {
+                SlimefunPlugin.getLocalization().sendMessage(p, "messages.enchantment-rune.fail", true);
             }
         }
     }